+++ lib/openmaize.ex
@@ -2,17 +2,19 @@ defmodule Openmaize do
   @moduledoc """
   Openmaize is an authentication / authorization library for Elixir.
 
-  It uses Plug extensively and provides the following main plugs:
+  It uses Plug extensively and provides the following plugs:
 
-  * Openmaize.LoginoutCheck
-      * checks the path to see if it is for the login or logout page
-      * handles login or logout if necessary
   * Openmaize.Authenticate
-      * authenticates the user
-      * sets (adds to the assigns map) the current_user variable
-
-  There are also plugs that can be used for authorization in the
-  Openmaize.AccessControl module.
+    * authenticates the user
+    * sets (adds to the assigns map) the current_user variable
+  * Openmaize.AccessControl.authorize
+    * check, based on the user's role, to see if the user is authorized to access the page
+  * Openmaize.AccessControl.authorize_id
+    * check, based on user id, to see if the user is authorized to access the page
+  * Openmaize.Login
+    * handle login POST request
+  * Openmaize.Logout
+    * handle logout request
 
   See the relevant module documentation for more details.