+++ test/login_test.exs
@@ -1,0 +1,5 @@
+defmodule Openmaize.LoginTest do
+  use ExUnit.Case
+  use Plug.Test
+
+end