+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
-## v0.10.2
+## v0.11.0
 
 * Enhancements
     * Added unique_id config value, so it's possible to use `email`, or anything else, instead of `name` to identify the user.
+    * Added check for confirmed email to the login.
+* Changes
+    * Replaced LoginoutCheck with Login and Logout plugs, so now no check is done for login / logout path.
 
 ## v0.10.0